Add JSON utilities with Jackson support
This commit introduces the JsonUtils class for JSON serialization and deserialization using the Jackson library, complete with support for Kotlin modules and Java 8 time module. Additionally, corresponding unit tests are added to ensure the correctness of the serialization and deserialization functionality. The Jackson library dependencies have been updated in the build configuration files to facilitate this new functionality.